os-specific-preparations: linux: add new boot_macOS.command.zip file, remove unneeded steps about brew asking for passwords
Some users may have difficulties running the boot shell script on the
Mac. This is a new script that the user can double-click and run in
Terminal.app, without the user needing to enter anything to run it
manually. It contains one additional line to cd into the containing
directory.

Homebrew does not ask for the user's password after install because it
sets permissions in a way where it's not needed, so the extra steps
mentioning "enter password" to install python and libusb are removed.