os-specific-preparations: linux: switch to user install with pip
Using pip with sudo to install global packages is a bad idea. Usually,
global packages should be managed by the distribution's package manager.
This uses --user to install to the user's home directory instead.

It also switches the command to use `python3 -m pip` to ensure the same
version is being used to install pyusb and run the script.