Change RCM guide to my rollback of Ghostboys guide.
Note that this does not violate GitHub TOS, GitHubs TOS gives anyone the right to fork and view a repository on GitHub.com .
I did not change anything in the guide itself. I merely forked and hard rolled back the repository from the JS redirecting commit and enabled GH
Pages.